+++ src/index.js
@@ -1,0 +1,3 @@
+export { analyzeProject, ANALYSIS_SCHEMA_VERSION } from "./analyzer.js";
+export { createGraphSnapshot, queryGraph, OPERATIONS } from "./graph-query.js";
+export { createMixedStructuralScannerBackend } from "./web-scanner.js";